move components
projects/create-project/samples/create-project-samples-tag.filter -> stacks/stack-details/select-template/create-project-samples-tag.filter

projects/project-details/project-details.controller -> /workspaces/workspace-details/workspace-projects/project-details/project-details.controller
projects/project-details/project-details.html -> /workspaces/workspace-details/workspace-projects/project-details/project-details.html
projects/project-details/project-details.styl -> /workspaces/workspace-details/workspace-projects/project-details/project-details.styl

projects/project-details/repository/project-repository-config/* -> /workspaces/workspace-details/workspace-projects/project-details/repository/project-repository-config/*

projects/list-projects/project-item/* -> /workspaces/workspace-details/workspace-projects/project-item/*
