+++ dashboard/src/app/workspaces/workspace-details/workspace-projects/project-details/project-details.html
