+++ dashboard/src/app/workspaces/workspace-details/workspace-projects/project-details/repository/project-repository.controller.ts
@@ -10,7 +10,7 @@
  */
 'use strict';
 
-import {gitMixinId, subversionMixinId} from '../repository/project-repository-data';
+import {gitMixinId, subversionMixinId} from './project-repository-data';
 
 export class ProjectRepositoryController {
