+++ dashboard/src/app/workspaces/workspace-details/workspace-projects/project-details/repository/project-repository.styl
