+++ dashboard/src/app/workspaces/workspace-details/workspace-projects/project-item/project-item.styl
